- SteganographyFilter improved (still work in progress) -> Better managing of  word wrapping - SaveAsDialog bug (removing original image, when overriding destination image, caused from config.Overwriting) - InvertCharEncryption improved (still work in progress)

+++ Picturez_Lib/BitmapWithTag.cs
@@ -120,16 +120,18 @@ namespace Picturez_Lib
 			else 
 				tmpNewFileName = config.Path + IOPath.DirectorySeparatorChar + relativeFileName;
 
-			if (config.FileOverwriting){
-				if (FileName != tmpNewFileName){
-					IOFile.Delete(FileName);
-				}
-			}else {
-				while(FileHelper.I.Exists(tmpNewFileName)){
-					tmpNewFileName = tmpNewFileName.Insert(tmpNewFileName.LastIndexOf('.'), "_1");				
+			if (config.StretchImage != ConvertMode.Editor)
+			{
+				if (config.FileOverwriting){
+					if (FileName != tmpNewFileName){
+						IOFile.Delete(FileName);
+					}
+				}else {
+					while(FileHelper.I.Exists(tmpNewFileName)){
+						tmpNewFileName = tmpNewFileName.Insert(tmpNewFileName.LastIndexOf('.'), "_1");				
+					}
 				}
 			}
-
 			FileName = tmpNewFileName;
 			#endregion FileOverwriting
